ts for router
