+++ package.json
@@ -71,6 +71,7 @@
     "tslint": "^5.10.0",
     "tslint-config-prettier": "^1.13.0",
     "tslint-eslint-rules": "^5.3.1",
+    "tslint-sonarts": "^1.7.0",
     "typescript": "^2.9.2"
   }
 }
