+++ src/routes/venues.ts
@@ -1,0 +1,24 @@
+import { Router } from 'express';
+
+const venues = Router();
+
+import {
+  create,
+  deleteId,
+  getAll,
+  getId,
+  updateId
+} from '../controllers/venue';
+
+venues
+  .route('/')
+  .get(getAll)
+  .post(create);
+
+venues
+  .route('/:id')
+  .get(getId)
+  .patch(updateId)
+  .delete(deleteId);
+
+export default venues;
